Remove Installer file only if it exists
The package doesn't include the 32-bit installer.

So, when Remove-Item -Force -ea 0 "$filePath32" is called, powershell 5.1 could error out, because it's checking the path of an empty variable.

This changes the package to remove the installer only when the filename variable is set and the path exists.

NOTE:
It's possible to instead simply remove everything that refers to 32-bit instead. But the way the PR changes things allows for the possibility if upstream were to bring back 32-bit support.